Renamed files and objects to `WebTerminal`
