+++ src/WebTerminal.css
@@ -4,7 +4,7 @@
   -moz-osx-font-smoothing: grayscale;
 }
 
-.VanillaTerm {
+.WebTerm {
   background-color: #222222;
   color: white;
   font-family: "Inconsolata", "Courier New", Courier, monospace;
@@ -13,22 +13,22 @@
   padding: 0.75em 1em;
 }
 
-.VanillaTerm .container {
+.WebTerm .container {
   width: 100%;
   height: 100%;
   overflow-y: scroll;
 }
-.VanillaTerm .container ::selection {
+.WebTerm .container ::selection {
   background-color: #1e496c;
   text-shadow: none !important;
 }
 
-.VanillaTerm output {
+.WebTerm output {
   clear: both;
   line-height: 1.25em;
   width: 100%;
 }
-.VanillaTerm output pre {
+.WebTerm output pre {
   white-space: pre-wrap;
   white-space: -moz-pre-wrap;
   white-space: -pre-wrap;
@@ -36,50 +36,50 @@
   word-wrap: break-word;
   margin: 0;
 }
-.VanillaTerm output span {
+.WebTerm output span {
   display: block;
   font-weight: normal;
   color: #dddddd;
 }
 
-.VanillaTerm .prompt,
-.VanillaTerm a {
+.WebTerm .prompt,
+.WebTerm a {
   color: #449dff;
   font-weight: bold;
 }
 
-.VanillaTerm .prompt {
+.WebTerm .prompt {
   line-height: 1.25em;
   margin-right: 0.5em;
   margin-top: 0.36em;
 }
 
-.VanillaTerm .command {
+.WebTerm .command {
   display: flex;
 }
 
-.VanillaTerm .command p {
+.WebTerm .command p {
   margin-top: 0.36em;
 }
 
-.VanillaTerm .command.input .prompt {
+.WebTerm .command.input .prompt {
   color: white;
   font-weight: normal;
 }
 
-.VanillaTerm .command.idle .input {
+.WebTerm .command.idle .input {
   display: none;
 }
 
-.VanillaTerm .command.hidden .prompt {
+.WebTerm .command.hidden .prompt {
   display: none;
 }
 
-.VanillaTerm .command.hidden .input {
+.WebTerm .command.hidden .input {
   max-width: 1px;
 }
 
-.VanillaTerm .command .input {
+.WebTerm .command .input {
   background-color: transparent;
   border: none;
   color: inherit;
@@ -91,7 +91,7 @@
   padding-top: 7px;
 }
 
-.VanillaTerm .command .spinner:before {
+.WebTerm .command .spinner:before {
   display: inline-block;
   content: "⠋";
   animation: spin 1s linear infinite;
